chore(codegen): gather integer operations to module integer
